+++ crd-generator/api/src/main/java/io/fabric8/crd/generator/annotation/SchemaSwap.java
@@ -17,15 +17,39 @@ package io.fabric8.crd.generator.annotation;
 
 import java.lang.annotation.*;
 
-/*
- * Used to tweak the behavior of the crd-generator
+/**
+ * Annotation that allows replacing a nested schema with one from another class.
+ *
+ * This is an alternative to {@link SchemaFrom} for cases when the classes
+ * are coming from an external source and fields cannot be annotated directly.
+ *
+ * @see SchemaFrom
  */
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.TYPE_USE, ElementType.TYPE })
 @Retention(RetentionPolicy.RUNTIME)
+@Repeatable(SchemaSwaps.class)
 public @interface SchemaSwap {
+  /**
+   * The owning class of the field whose type is to be replaced.
+   * <p>
+   * It is an error if the type is not used in the same schema hierarchy where the {@link SchemaSwap} is used.
+   */
   Class<?> originalType();
 
+  /**
+   * Name of the field whose type is to be replaced.
+   * <p>
+   * The name should be specified exactly as defined in the Java class, before any renames
+   * and transformations ({@code @JsonProperty} and similar) take place.
+   * <p>
+   * It is an error if the field does not exist on {@link #originalType()}
+   */
   String fieldName();
 
+  /**
+   * The replacement schema that will be used for the {@link #fieldName()} instead of its specified type
+   * <p>
+   * The default value of {@code void.class} causes the field to be skipped
+   */
   Class<?> targetType() default void.class;
 }