+++ crd-generator/test/src/test/java/io/fabric8/crd/generator/schemaswaps/SchemaSwapSpec.java
@@ -13,17 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.example.extraction;
+package io.fabric8.crd.generator.schemaswaps;
 
-import io.fabric8.crd.generator.annotation.PreserveUnknownFields;
-import io.fabric8.crd.generator.annotation.SchemaFrom;
+public class SchemaSwapSpec {
+  private SomeObject first;
+  private SomeObject second;
+  private AnotherObject third;
+  private YetAnotherObject fourth;
 
-public class ExtractionSpec {
+  static class SomeObject {
+    private int shouldBeString;
+  }
 
-  @SchemaFrom(type = FooExtractor.class)
-  private Foo foo;
+  static class AnotherObject {
+    private String shouldBeInt;
+  }
 
-  @PreserveUnknownFields
-  private Foo bar;
-
+  static class YetAnotherObject {
+    private String shouldBeSkipped;
+  }
 }