+++ java-generator/core/src/test/java/io/fabric8/java/generator/GeneratorTest.java
@@ -39,6 +39,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class GeneratorTest {
@@ -108,7 +109,13 @@ class GeneratorTest {
   @Test
   void testPrimitive() {
     // Arrange
-    JPrimitive primitive = new JPrimitive("test", defaultConfig, null, Boolean.FALSE, null);
+    JPrimitive primitive = new JPrimitive(
+        "test",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null,
+        null);
 
     // Act
     GeneratorResult res = primitive.generateJava();
@@ -119,10 +126,76 @@ class GeneratorTest {
   }
 
   @Test
+  void testPrimitiveWithNumericValidationConstraints() {
+    // Arrange
+    final Double expectedMaximum = 3.14;
+    final Double expectedMinimum = 0.0;
+    ValidationProperties validationProperties = ValidationProperties.Builder
+        .getInstance()
+        .withMaximum(expectedMaximum)
+        .withMinimum(expectedMinimum)
+        .build();
+    JPrimitive primitive = new JPrimitive(
+        "test",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null,
+        validationProperties);
+
+    // Act
+    GeneratorResult res = primitive.generateJava();
+
+    // Assert
+    assertEquals("test", primitive.getType());
+    assertEquals(0, res.getTopLevelClasses().size());
+    assertEquals(expectedMaximum, primitive.getMaximum());
+    assertEquals(expectedMinimum, primitive.getMinimum());
+    assertNull(primitive.getPattern());
+  }
+
+  @Test
+  void testPrimitiveWithAlphanumericValidationConstraints() {
+    // Arrange
+    final String expectedPattern = ".*whatever.*";
+    ValidationProperties validationProperties = ValidationProperties.Builder
+        .getInstance()
+        .withPattern(expectedPattern)
+        .build();
+    JPrimitive primitive = new JPrimitive(
+        "test",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null,
+        validationProperties);
+
+    // Act
+    GeneratorResult res = primitive.generateJava();
+
+    // Assert
+    assertEquals("test", primitive.getType());
+    assertEquals(0, res.getTopLevelClasses().size());
+    assertEquals(expectedPattern, primitive.getPattern());
+    assertNull(primitive.getMaximum());
+    assertNull(primitive.getMinimum());
+  }
+
+  @Test
   void testArrayOfPrimitives() {
     // Arrange
-    JArray array = new JArray(new JPrimitive("primitive", defaultConfig, null, Boolean.FALSE, null), defaultConfig, null,
-        Boolean.FALSE, null);
+    JArray array = new JArray(
+        new JPrimitive(
+            "primitive",
+            defaultConfig,
+            null,
+            Boolean.FALSE,
+            null,
+            null),
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = array.generateJava();
@@ -136,7 +209,13 @@ class GeneratorTest {
   void testMapOfPrimitives() {
     // Arrange
     JMap map = new JMap(
-        new JPrimitive("primitive", defaultConfig, null, Boolean.FALSE, null),
+        new JPrimitive(
+            "primitive",
+            defaultConfig,
+            null,
+            Boolean.FALSE,
+            null,
+            null),
         defaultConfig,
         null,
         Boolean.FALSE,
@@ -153,7 +232,17 @@ class GeneratorTest {
   @Test
   void testEmptyObject() {
     // Arrange
-    JObject obj = new JObject("v1alpha1", "t", null, null, false, "", "", defaultConfig, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        "v1alpha1",
+        "t",
+        null,
+        null,
+        false,
+        "",
+        "", defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -168,7 +257,18 @@ class GeneratorTest {
   void testEmptyObjectWithSuffix() {
     // Arrange
     Config config = new Config(null, null, Config.Suffix.ALWAYS, null, null, null);
-    JObject obj = new JObject("v1alpha1", "t", null, null, false, "", "MySuffix", config, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        "v1alpha1",
+        "t",
+        null,
+        null,
+        false,
+        "",
+        "MySuffix",
+        config,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -182,7 +282,18 @@ class GeneratorTest {
   @Test
   void testEmptyObjectWithoutNamespace() {
     // Arrange
-    JObject obj = new JObject(null, "t", null, null, false, "", "", defaultConfig, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        null,
+        "t",
+        null,
+        null,
+        false,
+        "",
+        "",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -200,7 +311,18 @@ class GeneratorTest {
     JSONSchemaProps newBool = new JSONSchemaProps();
     newBool.setType("boolean");
     props.put("o1", newBool);
-    JObject obj = new JObject("v1alpha1", "t", props, null, false, "", "", defaultConfig, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        "v1alpha1",
+        "t",
+        props,
+        null,
+        false,
+        "",
+        "",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -225,14 +347,25 @@ class GeneratorTest {
     props.put("o1", newBool);
     List<String> req = new ArrayList<>(1);
     req.add("o1");
-    JObject obj = new JObject("v1alpha1", "t", props, req, false, "", "", defaultConfig, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        "v1alpha1",
+        "t",
+        props,
+        req,
+        false,
+        "",
+        "",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
 
     // Assert
     Optional<ClassOrInterfaceDeclaration> clz = res.getTopLevelClasses().get(0).getCompilationUnit().getClassByName("T");
-    assertTrue(clz.get().getFieldByName("o1").get().getAnnotationByName("NotNull").isPresent());
+    assertTrue(clz.get().getFieldByName("o1").get().getAnnotationByName("Required").isPresent());
   }
 
   @Test
@@ -246,7 +379,13 @@ class GeneratorTest {
     enumValues.add(new TextNode("bar"));
     enumValues.add(new TextNode("baz"));
     props.put("e1", newEnum);
-    JEnum enu = new JEnum("t", enumValues, defaultConfig, null, Boolean.FALSE, null);
+    JEnum enu = new JEnum(
+        "t",
+        enumValues,
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = enu.generateJava();
@@ -276,7 +415,13 @@ class GeneratorTest {
     enumValues.add(new TextNode("bar"));
     enumValues.add(new TextNode("baz"));
     props.put("e1", newEnum);
-    JEnum enu = new JEnum("t", enumValues, new Config(false, null, null, null, null, null), null, Boolean.FALSE, null);
+    JEnum enu = new JEnum(
+        "t",
+        enumValues,
+        new Config(false, null, null, null, null, null),
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = enu.generateJava();
@@ -298,9 +443,22 @@ class GeneratorTest {
   void testArrayOfObjects() {
     // Arrange
     JArray array = new JArray(
-        new JObject(null, "t", null, null, false, "", "", defaultConfig, null, Boolean.FALSE, null),
+        new JObject(
+            null,
+            "t",
+            null,
+            null,
+            false,
+            "",
+            "",
+            defaultConfig,
+            null,
+            Boolean.FALSE,
+            null),
         defaultConfig,
-        null, false, null);
+        null,
+        false,
+        null);
 
     // Act
     GeneratorResult res = array.generateJava();
@@ -315,9 +473,22 @@ class GeneratorTest {
   void testMapOfObjects() {
     // Arrange
     JMap map = new JMap(
-        new JObject(null, "t", null, null, false, "", "", defaultConfig, null, Boolean.FALSE, null),
+        new JObject(
+            null,
+            "t",
+            null,
+            null,
+            false,
+            "",
+            "",
+            defaultConfig,
+            null,
+            Boolean.FALSE,
+            null),
         defaultConfig,
-        null, Boolean.FALSE, null);
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = map.generateJava();
@@ -335,7 +506,18 @@ class GeneratorTest {
     JSONSchemaProps newObj = new JSONSchemaProps();
     newObj.setType("object");
     props.put("o1", newObj);
-    JObject obj = new JObject(null, "t", props, null, false, "", "", defaultConfig, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        null,
+        "t",
+        props,
+        null,
+        false,
+        "",
+        "",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -361,7 +543,18 @@ class GeneratorTest {
     JSONSchemaProps newObj = new JSONSchemaProps();
     newObj.setType("object");
     props.put("o1", newObj);
-    JObject obj = new JObject(null, "t", props, null, false, "My", "", config, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        null,
+        "t",
+        props,
+        null,
+        false,
+        "My",
+        "",
+        config,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -380,7 +573,18 @@ class GeneratorTest {
     JSONSchemaProps newObj = new JSONSchemaProps();
     newObj.setType("object");
     props.put("o1", newObj);
-    JObject obj = new JObject(null, "t", props, null, false, "My", "", config, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        null,
+        "t",
+        props,
+        null,
+        false,
+        "My",
+        "",
+        config,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -394,7 +598,18 @@ class GeneratorTest {
   @Test
   void testObjectWithPreservedFields() {
     // Arrange
-    JObject obj = new JObject(null, "t", null, null, true, "", "", defaultConfig, null, Boolean.FALSE, null);
+    JObject obj = new JObject(
+        null,
+        "t",
+        null,
+        null,
+        true,
+        "",
+        "",
+        defaultConfig,
+        null,
+        Boolean.FALSE,
+        null);
 
     // Act
     GeneratorResult res = obj.generateJava();
@@ -439,20 +654,22 @@ class GeneratorTest {
     Optional<FieldDeclaration> o1Field = clzT.get().getFieldByName("o1");
     assertTrue(o1Field.isPresent());
     FieldDeclaration actualO1Field = o1Field.get();
-    Optional<AnnotationExpr> nullableAnnotation = actualO1Field
+    Optional<AnnotationExpr> nullableJacksonBasedAnnotation = actualO1Field
         .getAnnotationByName("com.fasterxml.jackson.annotation.JsonSetter");
-    assertTrue(nullableAnnotation.isPresent());
-    assertInstanceOf(SingleMemberAnnotationExpr.class, nullableAnnotation.get());
-    SingleMemberAnnotationExpr actualNullableAnnotation = (SingleMemberAnnotationExpr) nullableAnnotation.get();
+    assertTrue(nullableJacksonBasedAnnotation.isPresent());
+    assertInstanceOf(SingleMemberAnnotationExpr.class, nullableJacksonBasedAnnotation.get());
+    SingleMemberAnnotationExpr actualNullableAnnotation = (SingleMemberAnnotationExpr) nullableJacksonBasedAnnotation.get();
     assertEquals("nulls = com.fasterxml.jackson.annotation.Nulls.SET", actualNullableAnnotation.getMemberValue().toString());
+    Optional<AnnotationExpr> nullableFabric8BasedAnnotation = actualO1Field.getAnnotationByName("Nullable");
+    assertTrue(nullableFabric8BasedAnnotation.isPresent());
 
     Optional<FieldDeclaration> o2Field = clzT.get().getFieldByName("o2");
     assertTrue(o2Field.isPresent());
     FieldDeclaration actualO2Field = o2Field.get();
-    nullableAnnotation = actualO2Field.getAnnotationByName("com.fasterxml.jackson.annotation.JsonSetter");
-    assertTrue(nullableAnnotation.isPresent());
-    assertInstanceOf(SingleMemberAnnotationExpr.class, nullableAnnotation.get());
-    actualNullableAnnotation = (SingleMemberAnnotationExpr) nullableAnnotation.get();
+    nullableJacksonBasedAnnotation = actualO2Field.getAnnotationByName("com.fasterxml.jackson.annotation.JsonSetter");
+    assertTrue(nullableJacksonBasedAnnotation.isPresent());
+    assertInstanceOf(SingleMemberAnnotationExpr.class, nullableJacksonBasedAnnotation.get());
+    actualNullableAnnotation = (SingleMemberAnnotationExpr) nullableJacksonBasedAnnotation.get();
     assertEquals("nulls = com.fasterxml.jackson.annotation.Nulls.SKIP", actualNullableAnnotation.getMemberValue().toString());
 
     Optional<ClassOrInterfaceDeclaration> clzO1 = res.getTopLevelClasses().get(0).getCompilationUnit().getClassByName("O1");